Users {  read level access  write level access  edit level access  }
 File {
  read level access
  write level access
  edit level access
}

FS {
   Change dir name
   Change file name \ file accesses
}

+++ src/kernel/compilers/asm/asm_executor.c
@@ -59,7 +59,7 @@ int check_condition(int operand1, int operand2, int opcode){
 	return 0;
 }
 
-void asm_executor(int *memory_array, int memory_index){
+void asm_executor(int *memory_array, int memory_index, struct User* user) {
 	for (int i = 0; i < intermediate_index;){  // iterating on the intermediate language table
 
 		switch (intermediate_table[i]->opcode) {
@@ -91,23 +91,30 @@ void asm_executor(int *memory_array, int memory_index){
 	    	break;  // PRINTS Instruction //
 
 			case 18:
-				create_file(atoi(intermediate_table[i]->string_params[1]), 
+				create_file(atoi(intermediate_table[i]->string_params[1]), user->write_access, user->edit_access, 
 				intermediate_table[i]->string_params[0], ATA_find_empty_sector(200));
 			break;
 
 			case 19:
-				delete_file(intermediate_table[i]->string);
+				if (file_exist(intermediate_table[i]->string) == 1)
+					if (user->edit_access <= find_file(intermediate_table[i]->string)->edit_level)
+						delete_file(intermediate_table[i]->string);
 			break;
 
 			case 20:
 				struct File* wfile = find_file(intermediate_table[i]->string_params[0]);
-				write_file(wfile, intermediate_table[i]->string_params[1]);
+				if (wfile != NULL)
+					if (user->write_access <= wfile->write_level)
+						write_file(wfile, intermediate_table[i]->string_params[1]);
 			break;
 
 			case 21:
 				struct File* rfile = find_file(intermediate_table[i]->string);
-				int data = atoi(read_file(rfile));
-				memory_array[intermediate_table[i]->parameters[0]] = data; 
+				if (rfile != NULL)
+					if (user->read_access <= rfile->read_level) {
+						int data = atoi(read_file(rfile));
+						memory_array[intermediate_table[i]->parameters[0]] = data; 
+					}
 			break;
 
 			case 7:
